services/fetcher-service/requirements.txt
# FILE: ./services/fetcher-service/requirements.txt
requests
python-dotenv
python-json-logger
schedule
feedparser
pika
newspaper3k
lxml_html_clean
trafilatura